+++ src/shared/ui/alert/hooks/index.ts
