+++ src/shared/ui/alert/hooks/useAlert.tsx
